Update Any_Z_Wave_Dimmer_Driver_v1_5_5_425.groovy
There are some values that return as an empty string. instead of bool 0 or 1 for whether a device is read_only or not. 

When this happens it causes a `NumberFormatException` error. This try catch handles that and defaults to not read-only and allows the scanning and input creation to proceed.